+++ src/csr_matrix.h
@@ -4,9 +4,9 @@ typedef struct {
 	    l/*number of columns*/;
 	int *rows /*size is n*/;
 	int *cols;
-	double *vals /*rows values*/;
+	int *vals /*rows values*/;
 } sparse_matrix_t;
 
 void csr_create(sparse_matrix_t **csr_matrix, 
-		double **matrix,
+		int **matrix,
 		const int length, const int width);
